requirements.txt
appdirs==1.4.4
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.2
filelock==3.0.12
firebase-admin==4.5.0
Flask==2.0.1
gcloud==0.18.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jws==0.1.3
MarkupSafe==2.0.0
msgpack==1.0.2
mysql-connector-python==8.0.22
oauth2client==4.1.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
Pyrebase4==4.4.3
python-jwt==2.0.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.7
Werkzeug==2.0.1
